+++ R/install_workshops_pkgs.R
@@ -14,6 +14,9 @@ install_workshops_pkgs <- function(file = "pkgs.yaml", verbose = TRUE) {
   invisible(NULL)
 }
 
+#' @rdname install_workshops_pkgs
+#' @export
+
 install_workshop_pkgs_indiv <- function(file = "pkgs.yaml", verbose = TRUE) {
   ls_file <- read_yaml(file)
 
@@ -40,9 +43,10 @@ install_workshop_pkgs_indiv <- function(file = "pkgs.yaml", verbose = TRUE) {
 }
 
 
+
 install_cran_pkgs <- function(x) {
   for (i in seq_along(x)) {
-    if (!x[i] %in% rownames(installed.packages()))
+    if (!requireNamespace(x[i], quietly = TRUE))
       install.packages(x[i], repos = "https://cran.wu.ac.at/")
   }
   invisible(NULL)
@@ -50,7 +54,7 @@ install_cran_pkgs <- function(x) {
 
 install_cran_archived_pkgs <- function(x) {
   for (i in seq_along(x)) {
-    if (!x[[i]]$pkgname %in% rownames(installed.packages())) {
+    if (!requireNamespace(x[[i]]$pkgname, quietly = TRUE)) {
       f <- tempfile()
       download.file(x[[i]]$url, destfile = f)
       install.packages(pkgs = f, type="source", repos = NULL)
@@ -61,7 +65,7 @@ install_cran_archived_pkgs <- function(x) {
 
 install_repos_pkgs <- function(x) {
   for (i in seq_along(x)) {
-    if (!x[[i]]$pkgname %in% rownames(installed.packages())) {
+    if (!requireNamespace(x[[i]]$pkgname, quietly = TRUE)) {
        install.packages(x[[i]]$pkgname, repos = x[[i]]$repos, type="source")
     }
   }